生产跟踪单，活化组装人员

+++ eladmin-product/src/main/java/me/zhengjie/domain/ElectricPipeActivationDetail.java
@@ -42,7 +42,7 @@ public class ElectricPipeActivationDetail implements Serializable {
 
     @Column(name = "voltage")
     @ApiModelProperty(value = "总电压")
-    private Integer voltage;
+    private Double voltage;
 
     @Column(name = "power")
     @ApiModelProperty(value = "功率")
